feature: listando contas

+++ app/src/main/java/dev/thiaago/finn/features/home/domain/entities/AccountEntity.kt
@@ -3,14 +3,14 @@ package dev.thiaago.finn.features.home.domain.entities
 data class AccountEntity(
     val id: String? = null,
     val name: String,
-    val balance: Int = 0,
+    val balance: Long = 0,
 ) {
     companion object {
         fun fromMap(map: Map<String, Any>) : AccountEntity {
             return AccountEntity(
                 id = map["id"] as String?,
                 name = map["name"] as String? ?: "",
-                balance = map["balance"] as Int? ?: 0,
+                balance = map["balance"] as Long? ?: 0,
             )
         }
     }
